Move contextvars into separate module
Goal is to clean up the toplevel module to contain only external API.
